library/models.py spread to models dir
