+++ apps/library/models/participation_application.py
@@ -1,0 +1,95 @@
+from django.core.validators import FileExtensionValidator
+from django.db import models
+from phonenumber_field.modelfields import PhoneNumberField
+
+from apps.core.models import BaseModel
+from apps.core.utilities import slugify
+from apps.library.utilities import generate_class_name_path
+from apps.library.validators import year_validator
+
+
+class ParticipationApplicationFestival(BaseModel):
+    """
+    Заявки на участие в фестивале
+    """
+
+    first_name = models.CharField(
+        max_length=50,
+        verbose_name="Имя",
+    )
+    last_name = models.CharField(
+        max_length=50,
+        verbose_name="Фамилия",
+    )
+    birthday = models.DateField(
+        verbose_name="День рождения",
+    )
+    city = models.CharField(
+        max_length=50,
+        verbose_name="Город проживания",
+    )
+    phone_number = PhoneNumberField()
+    email = models.EmailField(
+        max_length=100,
+        verbose_name="Электронная почта",
+    )
+    title = models.CharField(
+        max_length=200,
+        verbose_name="Название пьесы",
+    )
+    year = models.PositiveSmallIntegerField(
+        validators=(year_validator,),
+        verbose_name="Год написания",
+    )
+    file = models.FileField(
+        validators=(
+            FileExtensionValidator(["doc", "docx", "txt", "odt", "pdf"]),
+        ),
+        verbose_name="Файл",
+        upload_to=generate_class_name_path,
+    )
+
+    BOOL_CHOICES = ((True, "Да"), (False, "Нет"))
+    verified = models.BooleanField(
+        default=False,
+        verbose_name="Проверена?",
+        choices=BOOL_CHOICES,
+    )
+
+    class Meta:
+        verbose_name_plural = "Заявления на участие"
+        verbose_name = "Заявление на участие"
+        constraints = (
+            models.UniqueConstraint(
+                fields=(
+                    "first_name",
+                    "last_name",
+                    "birthday",
+                    "city",
+                    "phone_number",
+                    "email",
+                    "title",
+                    "year",
+                ),
+                name="unique_application",
+            ),
+        )
+
+    def __str__(self):
+        return f"{self.last_name}-{self.title}"
+
+    def generate_filename(self):
+        """
+        Generate new filename as "Last_name-Title" format
+        """
+
+        filename = f"{self.last_name}_{self.first_name}___{self.title}"
+        filename = slugify(filename).replace("-", "_")
+        return f"{filename.title()}.{self.file.name.split('.')[1]}"
+
+    def save(self, *args, **kwargs):
+        """
+        Save generated filename
+        """
+        self.file.name = self.generate_filename()
+        super().save(*args, **kwargs)
